// src/state-management/create-tab/enums.ts
export enum CreateTabActionType {
  SetTabName = 'CreateTab/SetTabName',
  SetVenue = 'CreateTab/SetVenue',
  SetCurrency = 'CreateTab/SetCurrency',
  SetNotes = 'CreateTab/SetNotes',
  AddMember = 'CreateTab/AddMember',
  RemoveMember = 'CreateTab/RemoveMember',
  AddMenuItem = 'CreateTab/AddMenuItem',
  RemoveMenuItem = 'CreateTab/RemoveMenuItem',
  Reset = 'CreateTab/Reset',
  SubmitPending = 'CreateTab/SubmitPending',
  SubmitSuccess = 'CreateTab/SubmitSuccess',
  SubmitFailure = 'CreateTab/SubmitFailure',
}